+++ legadilo/users/tests/models/test_user.py
@@ -10,9 +10,12 @@ from allauth.account.models import EmailAddress
 from django.core.exceptions import ValidationError
 from django.core.management import call_command
 from django.db import IntegrityError
+from slugify import slugify
 
+from legadilo.core.utils.testing import AnyOfType
 from legadilo.core.utils.time_utils import utcnow
 from legadilo.feeds.tests.factories import FeedFactory
+from legadilo.users import constants
 from legadilo.users.models import User, UserSession
 from legadilo.users.tests.factories import UserFactory
 
@@ -88,6 +91,63 @@ class TestUserQuerySet:
             user_to_delete_inactive_and_no_verified_email,
             user_to_delete_active_and_no_verified_email,
         ]
+
+    def test_inactive_accounts_to_delete(self):
+        UserFactory(email="active@example.com", last_login=utcnow())
+        user_to_delete = UserFactory(
+            email="to_delete@example.com", last_login=utcnow() - constants.INACTIVE_USERS_RETENTION
+        )
+        UserFactory(
+            email="to_notify@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[0],
+        )
+
+        inactive_accounts = User.objects.get_queryset().inactive_accounts_to_delete()
+
+        assert list(inactive_accounts) == [user_to_delete]
+
+    def test_inactive_accounts_to_notify(self):
+        UserFactory(email="active@example.com", last_login=utcnow())
+        UserFactory(
+            email="to_delete@example.com", last_login=utcnow() - constants.INACTIVE_USERS_RETENTION
+        )
+        user_to_notify_30_days = UserFactory(
+            email="user_to_notify_30_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[0],
+        )
+        user_to_notify_14_days = UserFactory(
+            email="user_to_notify_14_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[1],
+        )
+        user_to_notify_7_days = UserFactory(
+            email="user_to_notify_7_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[2],
+        )
+        UserFactory(
+            email="almost_to_notify@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[0]
+            + timedelta(days=1),
+        )
+
+        inactive_account_ids = (
+            User.objects.get_queryset().inactive_accounts_to_notify().values_list("id", flat=True)
+        )
+
+        assert set(inactive_account_ids) == {
+            user_to_notify_30_days.id,
+            user_to_notify_14_days.id,
+            user_to_notify_7_days.id,
+        }
 
 
 @pytest.mark.django_db
@@ -223,6 +283,70 @@ class TestUserManager:
         admin_emails = User.objects.list_admin_emails()
 
         assert admin_emails == [admin_user.email]
+
+    def test_notify_inactive_accounts(self, mocker, snapshot):
+        send_mail_mock = mocker.patch("legadilo.users.models.user.send_mail")
+        user_to_notify_14_days = UserFactory(
+            email="user_to_notify_14_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[1],
+        )
+        user_to_notify_7_days = UserFactory(
+            email="user_to_notify_7_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[2],
+        )
+        UserFactory(
+            email="almost_to_notify@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[0]
+            + timedelta(days=1),
+        )
+
+        nb_notified_accounts = User.objects.notify_inactive_accounts()
+
+        assert nb_notified_accounts == 2
+        send_mail_mock.assert_has_calls(
+            [
+                mocker.call(
+                    subject="Your Legadilo account is inactive and will be deleted soon",
+                    message=AnyOfType(str),
+                    from_email="Legadilo <noreply@legadilo.eu>",
+                    recipient_list=[user_to_notify_7_days.email],
+                ),
+                mocker.call(
+                    subject="Your Legadilo account is inactive and will be deleted soon",
+                    message=AnyOfType(str),
+                    from_email="Legadilo <noreply@legadilo.eu>",
+                    recipient_list=[user_to_notify_14_days.email],
+                ),
+            ],
+            any_order=True,
+        )
+        for call in send_mail_mock.call_args_list:
+            email = call.kwargs["recipient_list"][0]
+            snapshot.assert_match(call.kwargs["message"], f"message_body_{slugify(email)}.txt")
+
+    def test_cleanup_inactive_users(self):
+        user_active = UserFactory(email="active@example.com", last_login=utcnow())
+        UserFactory(
+            email="to_delete@example.com", last_login=utcnow() - constants.INACTIVE_USERS_RETENTION
+        )
+        user_to_notify_14_days = UserFactory(
+            email="user_to_notify_14_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[1],
+        )
+
+        nb_deleted = User.objects.cleanup_inactive_users()
+
+        assert nb_deleted == (2, {"users.UserSettings": 1, "users.User": 1})
+        user_ids = set(User.objects.values_list("id", flat=True))
+        assert user_ids == {user_active.id, user_to_notify_14_days.id}
 
 
 @pytest.mark.django_db