+++ legadilo/users/tests/test_commands/test_clean_users.py
@@ -9,6 +9,7 @@ from allauth.account.models import EmailAddress
 from django.core.management import call_command
 
 from legadilo.core.utils.time_utils import utcnow
+from legadilo.users import constants
 from legadilo.users.models import User, UserSession
 from legadilo.users.tests.factories import UserFactory
 
@@ -49,3 +50,20 @@ class TestCleanUsersCommand:
         assert UserSession.objects.count() == 1
         session = UserSession.objects.get()
         assert session.session_key == "valid_session"
+
+    def test_clean_inactive_users(self, mocker):
+        send_mail_mock = mocker.patch("legadilo.users.models.user.send_mail")
+        UserFactory(
+            email="user_to_notify_14_days@example.com",
+            last_login=utcnow()
+            - constants.INACTIVE_USERS_RETENTION
+            + constants.INACTIVE_USERS_NOTIFICATION_THRESHOLDS[1],
+        )
+        UserFactory(
+            email="to_delete@example.com", last_login=utcnow() - constants.INACTIVE_USERS_RETENTION
+        )
+
+        call_command("clean_users")
+
+        assert send_mail_mock.call_count == 1
+        assert User.objects.count() == 1